Improve reliability of the validation plugin
Because we can't rely on form element, it appears that the key used in $.data(key…) may mismatch; thus, causing validator data not to be found. Thus, the fix is to cache the key in the target form itself.

Also, reversed the key evaluation. Instead of looking for native DOM form element, we will retrieve the cached first via jQuery first.
This is to cater special case where you are validating a confined group of input field inside a global form element.